requirements.txt
pandas>=1.5
numpy
matplotlib
scikit-learn
python-dateutil
openpyxl
rapidfuzz     # fuzzy string matching (faster than fuzzywuzzy)
pyyaml